+++ OP2Utility.vcxproj
@@ -199,10 +199,10 @@
     <ClInclude Include="src\Stream\BidirectionalWriter.h" />
     <ClInclude Include="src\Stream\Reader.h" />
     <ClInclude Include="src\Stream\Writer.h" />
-    <ClInclude Include="src\StringHelper.h" />
+    <ClInclude Include="src\StringUtility.h" />
     <ClInclude Include="src\Tag.h" />
     <ClInclude Include="src\XFile.h" />
-    <ClCompile Include="src\StringHelper.cpp" />
+    <ClCompile Include="src\StringUtility.cpp" />
     <ClCompile Include="src\XFile.cpp" />
     <ClInclude Include="src\Sprite\TilesetLoader.h" />
   </ItemGroup>
