+++ src/StringUtility.h
